requirements.txt
aniso8601==3.0.2
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pytest-cache==1.0
pytest-cov==2.5.1
pytest-pep8==1.0.6
pytest-flask==0.10.0
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
